[TASK] Move the repository unit tests to functional
This saves us the work of instantiating the repositories and their
dependencies ourselves in the tests.
