+++ Tests/Functional/Domain/Repository/FrontendUserGroupRepositoryTest.php
@@ -5,8 +5,11 @@ declare(strict_types=1);
 namespace OliverKlee\FeUserExtraFields\Tests\Functional\Domain\Repository;
 
 use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUserGroup;
+use OliverKlee\FeUserExtraFields\Domain\Repository\DirectPersistInterface;
 use OliverKlee\FeUserExtraFields\Domain\Repository\FrontendUserGroupRepository;
 use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
+use TYPO3\CMS\Extbase\Persistence\Repository;
+use TYPO3\CMS\Extbase\Persistence\RepositoryInterface;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
@@ -24,6 +27,30 @@ final class FrontendUserGroupRepositoryTest extends FunctionalTestCase
         parent::setUp();
 
         $this->subject = $this->get(FrontendUserGroupRepository::class);
+    }
+
+    /**
+     * @test
+     */
+    public function implementsRepositoryInterface(): void
+    {
+        self::assertInstanceOf(RepositoryInterface::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function isRepository(): void
+    {
+        self::assertInstanceOf(Repository::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function implementsDirectPersistInterface(): void
+    {
+        self::assertInstanceOf(DirectPersistInterface::class, $this->subject);
     }
 
     /**
